requirements.txt
fastapi
uvicorn
pydantic
dependency_injector
supabase
python-dotenv
sqlalchemy
psycopg2-binary
datetime
email-validator